Manage state in one place
Every time I start a React project I have to be reminded, at some point, that components update in response to changes in state. Thus, you should be managing state in one place. If you have a data structure as part of your state, you need to manage it in one component (or in Redux / Context, etc.). This refactor moves state management into the Recorder component. We’re still a little shy on tests, but Edit and Delete buttons are enabled.