Visualizer works but…
The visualizer is working nicely but the code itself is not terribly compliant with the React way of doing things. There are scope dependencies (canvas and canvasCtx set in useEffect), a function inside another function, and I suspect there are renders happening that are interfering with the painting on the canvas, or vice versa (the canvas should only be rendered *once*).