Bug fix: test erroneously reports duplicates error
Fixed an issue in which one of the tests was reporting duplicate IDs on
the list of recordings in the MOCKED version. This is the exact reason
why you should mock as little as possible. This was not happening in the
actual application because the ID generation code was much better.