Issue #6 Cleanup before refactor adding R Router
This is just a little cleanup before adding React Router. The aim going
forward is to keep the tests the same while using React Router to
generate the A element for Terms of Use. That's the expected outcome:
Using a Link component, I end up with an A element that matches my tests