Combined two steps in one
I was unaware that destructering was the same as initializing so my workaround for two variables with the same name (assigning with let vs. const) didn't work and I had to simply delete the let assignment and use the destructuring assignment. Again, here, I broke my rule of only changing one line but they are literally right next to each other so I'm not bothered by it.