Flaky test: tests pass when render is explicit
When I try to use recorderRenderer in Recorder, the tests all turn red. I'm double-checking the tests to make sure they are solid, and at first glance they are not - I've had to do a little clean up and the mocking is being flagged by TS as incorrect. I'm trying _not_ to change anything in the application files until I'm certain the Recorder tests are valid.