New Refactor Approach - RecorderProvider
Changing the refactoring approach. Instead of using custom hooks, which
don't really work in this scenario, I'm using a "provider" that
initializes the MediaRecorder object and passes it to the Recorder
component as a property. This should allow me to get decent test
coverage and easily mock the native browser objects.