FE: increase evidence there are audios below the fold
Currently, if you add numerous audios, they get added to the bottom of
the list and you don't see them appear. This is an attempt to improve
visibility although I think the ultimate solution will be to add new
recordings to the top of the list rather than the bottom.