Shameless Green
In the truest sense of the word, this is truly shameless.

I'm having serious problems mocking the MediaRecorder object and I
suspect it is due to some sort of name collision as the MediaRecorder
has a member named 'state', which is also some kind of internal React
reserved word… Thus, I am completely unable to modify its value when
running tests. I have lost HOURS trying to get this to work but finally
gave up and will come back to it later.